+++ noodles-vcf/src/header/format.rs
@@ -1,3 +1,0 @@
-//! VCF header genotype format record components.
-
-pub mod key;