+++ noodles-vcf/src/record/builder.rs
@@ -274,7 +274,6 @@ impl Builder {
     /// ```
     /// use noodles_vcf::{
     ///     self as vcf,
-    ///     header::format::key,
     ///     record::{genotypes::sample::Value, Genotypes, Position},
     /// };
     ///